making RALibretro cross-compilable
when trying to cross-compile RALibretro on a Linux machine, the compiler complains about these header files. It's just a matter of "case-sensitiveness".

Unfortunately I'm unable to compile RA_Integration and I am not 100% sure that the changes here would cause some problem. But, as Windows isn't case sensitive, I think it's fine.

Can some of you guys test it?